+++ src/Codeception/Command/Bootstrap.php
@@ -10,6 +10,7 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Yaml\Yaml;
+use Symfony\Component\Console\Question\Question;
 
 /**
  * Creates default config, tests directory and sample suites for current project. Use this command to start building a test suite.
@@ -78,21 +79,13 @@ class Bootstrap extends Command
         $output->writeln(
             "<fg=white;bg=magenta> Initializing Codeception in " . $realpath . " </fg=white;bg=magenta>\n"
         );
-
-        $this->createGlobalConfig();
-        $output->writeln("File codeception.yml created       <- global configuration");
-        $this->createDirs();
-
-        if (!$input->getOption('empty')) {
-            $this->createUnitSuite();
-            $output->writeln("tests/unit created                 <- unit tests");
-            $output->writeln("tests/unit.suite.yml written       <- unit tests suite configuration");
-            $this->createFunctionalSuite();
-            $output->writeln("tests/functional created           <- functional tests");
-            $output->writeln("tests/functional.suite.yml written <- functional tests suite configuration");
-            $this->createAcceptanceSuite();
-            $output->writeln("tests/acceptance created           <- acceptance tests");
-            $output->writeln("tests/acceptance.suite.yml written <- acceptance tests suite configuration");
+        
+        if ($input->getOption('compat')) {
+            $this->compatibilitySetup($output);
+        } elseif ($input->getOption('customize')) {
+            $this->customize($input, $output);
+        } else {
+            $this->setup($output);
         }
         $output->writeln(" --- ");
         $this->ignoreFolderContent('tests/_output');