+++ src/Codeception/Command/SelfUpdate.php
@@ -2,6 +2,7 @@
 namespace Codeception\Command;
 
 use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Question\ConfirmationQuestion;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
@@ -75,12 +76,9 @@ class SelfUpdate extends Command
                 );
                 if (!$input->getOption('no-interaction')) {
 
-                    $dialog = $this->getHelperSet()->get('dialog');
-                    if (!$dialog->askConfirmation(
-                        $output,
-                        "\n<question>Do you want to update?</question> ", false
-                    )
-                    ) {
+                    $dialog = $this->getHelperSet()->get('question');
+                    $question = new ConfirmationQuestion("\n<question>Do you want to update?</question> ", false);
+                    if (!$dialog->ask($input, $output, $question)) {
                         $output->writeln("\n<info>Bye-bye!</info>\n");
 
                         return;