+++ src/Codeception/Lib/Connector/ZF1.php
@@ -57,7 +57,7 @@ class ZF1 extends Client
         // relying on $request->getPost()
         $zendRequest->setPost($request->getParameters());
         $zendRequest->setRequestUri(str_replace('http://localhost','',$request->getUri()));
-        $zendRequest->setHeaders($request->getServer());
+        $zendRequest->setHeaders($this->extractHeaders($request));
         $_FILES  = $this->remapFiles($request->getFiles());
         $_SERVER = array_merge($_SERVER, $request->getServer());
 
@@ -78,7 +78,7 @@ class ZF1 extends Client
         $response = new Response(
             $zendResponse->getBody(),
             $zendResponse->getHttpResponseCode(),
-            $this->_formatResponseHeaders($zendResponse)
+            $this->formatResponseHeaders($zendResponse)
         );
 
         return $response;
@@ -90,9 +90,8 @@ class ZF1 extends Client
      * @param \Zend_Controller_Response_Abstract $response The ZF1 Response Object.
      * @return array the clean key/value headers
      */
-    private function _formatResponseHeaders(\Zend_Controller_Response_Abstract $response)
-    {
-        $headers = [];
+    private function formatResponseHeaders (\Zend_Controller_Response_Abstract $response) {
+        $headers = array();
         foreach ($response->getHeaders() as $header) {
             $name = $header['name'];
             if (array_key_exists($name, $headers)) {
@@ -115,7 +114,7 @@ class ZF1 extends Client
         return $this->zendRequest;
     }
 
-    private function _extractHeaders(BrowserKitRequest $request)
+    private function extractHeaders(BrowserKitRequest $request)
     {
         $headers = array();
         $server = $request->getServer();
@@ -126,7 +125,7 @@ class ZF1 extends Client
             $server['HTTP_HOST'] .= ':' . $port;
         }
 
-        $contentHeaders = array('Content-length' => true, 'Content-md5' => true, 'Content-type' => true);
+        $contentHeaders = array('Content-Length' => true, 'Content-Md5' => true, 'Content-Type' => true);
         foreach ($server as $header => $val) {
             $header = implode('-', array_map('ucfirst', explode('-', strtolower(str_replace('_', '-', $header)))));