+++ src/Codeception/Lib/Connector/ZF2.php
@@ -66,7 +66,7 @@ class ZF2 extends Client
         $zendRequest->setUri($uri);
         $zendRequest->setRequestUri(str_replace('http://localhost','',$request->getUri()));
         
-        $zendRequest->setHeaders($this->_extractHeaders($request));
+        $zendRequest->setHeaders($this->extractHeaders($request));
         $this->application->run();
 
         $this->zendRequest = $zendRequest;
@@ -93,7 +93,7 @@ class ZF2 extends Client
         return $this->zendRequest;
     }
 
-    private function _extractHeaders(BrowserKitRequest $request)
+    private function extractHeaders(BrowserKitRequest $request)
     {
         $headers = array();
         $server = $request->getServer();
@@ -104,7 +104,7 @@ class ZF2 extends Client
             $server['HTTP_HOST'] .= ':' . $port;
         }
 
-        $contentHeaders = array('Content-length' => true, 'Content-md5' => true, 'Content-type' => true);
+        $contentHeaders = array('Content-Length' => true, 'Content-Md5' => true, 'Content-Type' => true);
         foreach ($server as $header => $val) {
             $header = implode('-', array_map('ucfirst', explode('-', strtolower(str_replace('_', '-', $header)))));