+++ tests/data/claypit/tests/dummy/AnotherCest.php
@@ -1,5 +1,4 @@
 <?php
-use \DumbGuy;
 
 class AnotherCest
 {